feat(e2e): move login commands to tools
