+++ apps/metadata-editor-e2e/src/e2e/home.cy.ts
@@ -11,7 +11,7 @@ const gnBaseUrl = 'http://localhost:8080/geonetwork/srv/eng/'
 describe('avatar', () => {
   describe('display avatar for user without gravatar hash', () => {
     it('should display placeholder url', () => {
-      cy.loginGN('admin', 'admin', false)
+      cy.login('admin', 'admin', false)
       cy.visit(`${gnBaseUrl}admin.console#/organization`)
       cy.get('#gn-btn-user-add').click()
       cy.get('#username').type(fakeUser.username)
@@ -33,7 +33,7 @@ describe('avatar', () => {
         .should('eq', 'https://www.gravatar.com/avatar/?d=mp')
     })
     it('should display monsterid', () => {
-      cy.loginGN('admin', 'admin', false)
+      cy.login('admin', 'admin', false)
       cy.visit(`${gnBaseUrl}admin.console#/settings`)
       cy.get('[id="system/users/identicon"]').type(
         '{selectAll}gravatar:monsterid'
@@ -49,7 +49,7 @@ describe('avatar', () => {
   })
   describe('display avatar for user with hash', () => {
     it('should display the correct profile picture', () => {
-      cy.loginGN(fakeUser.username, fakeUser.password)
+      cy.login(fakeUser.username, fakeUser.password)
       cy.get('gn-ui-avatar')
         .children('img')
         .should('have.attr', 'src')
