+++ tools/e2e/README.md
@@ -1,0 +1,4 @@
+# End-to-end testing utils
+
+This folder contains useful Cypress commands to be used in the E2E tests of
+various applications.
